Vastly improved S3M support, thanks to Opus 4.7

+++ SharpMod/Helpers/ExtensionMethods.cs
@@ -11,9 +11,9 @@ namespace SharpMod.Helpers {
             return BitConverter.ToUInt16(tmp2, 0);
         }
 
-        public static UInt16 ReadUInt32(this FileStream fs) {
+        public static UInt32 ReadUInt32(this FileStream fs) {
             fs.Read(tmp4, 0, 4);
-            return BitConverter.ToUInt16(tmp4, 0);
+            return BitConverter.ToUInt32(tmp4, 0);
         }
     }
 }
